moved createQuestLink function to its own file
