+++ map.html
@@ -20,11 +20,11 @@
     </header>
     <main>
         <h2>Choose Your World</h2>
-        <nav id="quests"></nav>
+        
         <div id="tunnels">
-           <a class="tunnel-link" href="quest.html"><img src="./assets/tunnel.png"></a>
+           <!-- <a class="tunnel-link" href="quest.html"><img src="./assets/tunnel.png"></a>
             <img class="tunnel-link" src="./assets/tunnel.png">
-            <img class="tunnel-link" src="./assets/tunnel.png">
+            <img class="tunnel-link" src="./assets/tunnel.png"> -->
         </div>
     </main>
