+++ test/create-quest-link.test.js
@@ -1,21 +1,7 @@
+import createQuestLink from '../src/map/create-quest-link.js';
 const test = QUnit.test;
 
 QUnit.module('create quest link');
-
-function createQuestLink(quest) {
-    const link = document.createElement('a');
-    link.classList.add('tunnel-link');
-    const image = document.createElement('img');
-    image.setAttribute('src', './assets/tunnel.png');
-    link.appendChild(image);
-
-    const searchParams = new URLSearchParams();
-    searchParams.set('id', quest.id);
-    link.href = 'quest.html?' + searchParams.toString();
-
-    return link;
-
-}
 
 test('create correct html', (assert) => {
     //arrange
